pdf_cover_page module for those functions.
